requirements.txt
enum34==1.1.10
GPUtil==1.4.0
netifaces==0.11.0
psutil==5.9.0
pyserial==3.5
pyspectator==1.2.2
pyvalid==1.0.4
pywin32==303
PyYAML==6.0
six==1.16.0
WMI==1.5.1
